requirements.txt
crewai==0.175.0
fastapi==0.116.1
langchain_community==0.3.29
langchain_openai==0.3.32
pydantic==2.11.7
python-dotenv==1.1.1